Fix api and web routes
