+++ settings/routes.php
@@ -148,6 +148,10 @@ function addWebRoutes(RouterService $routerService, FastRoute\RouteCollector $ro
     $routes->add('DELETE', '/settings/users/{userId:\d+}', [Web\UserController::class, 'deleteUser'], [Web\Middleware\UserIsAuthenticated::class]);
 
     $routes->add('GET', '/settings/integrations/radarr', [Web\SettingsController::class, 'renderRadarrPage'], [Web\Middleware\UserIsAuthenticated::class]);
+    $routes->add('PUT', '/settings/radarr/feed', [RadarrController::class, 'regenerateRadarrFeedUrl'], [Web\Middleware\UserIsAuthenticated::class]);
+    $routes->add('DELETE', '/settings/radarr/feed', [RadarrController::class, 'deleteRadarrFeedUrl'], [Web\Middleware\UserIsAuthenticated::class]);
+
+
     ##########
     # Movies #
     ##########
@@ -210,10 +214,9 @@ function addApiRoutes(RouterService $routerService, FastRoute\RouteCollector $ro
     $routes->add('DELETE', $routeUserPlayed, [Api\PlayedController::class, 'deleteFromPlayed'], [Api\Middleware\IsAuthorizedToWriteUserData::class]);
     $routes->add('PUT', $routeUserPlayed, [Api\PlayedController::class, 'updatePlayed'], [Api\Middleware\IsAuthorizedToWriteUserData::class]);
 
-    $routes->add('GET', '/radarr/{id:.+}', [RadarrController::class, 'renderRadarrFeed']);
-    $routes->add('PUT', '/radarr/regeneratefeedid', [RadarrController::class, 'regenerateRadarrFeedId'], [Web\Middleware\UserIsAuthenticated::class]);
-    $routes->add('DELETE', '/radarr/deletefeedid', [RadarrController::class, 'deleteRadarrFeedId'], [Web\Middleware\UserIsAuthenticated::class]);
     $routes->add('GET', '/movies/search', [Api\MovieSearchController::class, 'search'], [Api\Middleware\IsAuthenticated::class]);
+
+    $routes->add('GET', '/feed/radarr/{id:.+}', [Api\RadarrController::class, 'renderRadarrFeed']);
 
     $routerService->addRoutesToRouteCollector($routeCollector, $routes);
 }
